Split tests between four files to optimize test up times.
